+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod args;
 pub mod commands;
-pub mod error;
 pub mod logging;
 
 mod client;