+++ src/orchestration.rs
@@ -1,10 +1,12 @@
+use core::fmt;
+use std::fmt::Display;
+
 use log::*;
 
-use crate::client::WebClient;
+use crate::client::{WebClient, WebGetError};
 use crate::date::DateId;
-use crate::db::DatabaseHandle;
+use crate::db::{DatabaseGetError, DatabaseHandle, DatabaseInitError};
 use crate::display::DisplaySettings;
-use crate::error::{DatabaseError, DbUpdateError, RetrievalError};
 use crate::lectionary::Lectionary;
 
 /// Retrieves lectionary from db and web and attempts to store it before printing to STDOUT
@@ -97,4 +99,67 @@ async fn retrieve_for_database(date_id: DateId, db: &DatabaseHandle, client: &We
         }
         Err(web_error) => Err(web_error.into()),
     }
+}
+
+/// A failure to retrieve a lectionary from the database, web, or both
+///
+/// Used when trying to display a Lectionary
+#[derive(thiserror::Error, Debug)]
+pub struct RetrievalError {
+    db_error: Option<DatabaseError>,
+    web_error: Option<WebGetError>,
+}
+
+impl Display for RetrievalError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match (self.db_error.as_ref(), self.web_error.as_ref()) {
+            (Some(db_error), Some(web_error)) => write!(
+                f,
+                "Failed to retrieve from db ({db_error}) and failed to retrieve from web ({web_error})"
+            ),
+            (None, Some(web_error)) => write!(f, "Failed to retrieve from web ({web_error})"),
+            (Some(db_error), None) => write!(f, "Failed to retrieve from db ({db_error})"),
+            (None, None) => write!(f, "Failed to retrieve (undertermined cause)"),
+        }
+    }
+}
+
+impl From<DatabaseError> for RetrievalError {
+    fn from(value: DatabaseError) -> Self {
+        RetrievalError {
+            db_error: Some(value),
+            web_error: None,
+        }
+    }
+}
+
+impl From<WebGetError> for RetrievalError {
+    fn from(value: WebGetError) -> Self {
+        RetrievalError {
+            db_error: None,
+            web_error: Some(value),
+        }
+    }
+}
+
+/// Represents a failure to update the database with new entries
+#[derive(thiserror::Error, Debug)]
+pub enum DbUpdateError {
+    /// Failure to retrieve from Web
+    #[error("Could not retrieve lectionary for storage: ({0})")]
+    RetrieveError(#[from] WebGetError),
+    /// Failure to insert the retrieved lectionary in to database
+    #[error("Could not store retrieved lectionary in database: ({0})")]
+    InsertError(#[from] sqlx::Error),
+}
+
+/// Represents a critical failure while working with the database
+#[derive(thiserror::Error, Debug)]
+pub enum DatabaseError {
+    #[error("{0}")]
+    InitError(#[from] DatabaseInitError),
+    #[error("{0}")]
+    GetError(#[from] DatabaseGetError),
+    #[error("Failed to delete row(s) from the database: ({0})")]
+    DeleteError(#[from] sqlx::Error),
 }