+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - More Logging
 - Updated dependencies
     - scraper to 0.22.0
+- Added thiserror as dependency to refactor Error types
 
 ## [0.3.2] - 2024-12-09
 - Code lint